Create rent_contract
This smart contract, written in the Solidity software language, runs on the Ethereum network and ensures that rents are paid by wallets on the blockchain network. It also includes the smart contract to automatically take the necessary actions in case of non-payment of rent, over-increase, unfair termination of the contract.